# Use IsKind instead of type checking in syntax normalizer (little performance gain)