There seem to be a *very little* performance gain here:

**Before first commit (run on 3.8.0):**

|               Method |      Mean |    Error |   StdDev |    Median |       Min |       Max | Gen 0 |     Gen 1 | Gen 2 | Allocated |
|--------------------- |----------:|---------:|---------:|----------:|----------:|----------:|-----------:|----------:|------:|----------:|
|  NormalizeWhitespace | 815.74 ms | 7.782 ms | 6.898 ms | 814.23 ms | 807.54 ms | 832.21 ms | 25000.0000 | 7000.0000 |     - |    146 MB |



**After first commit:**

|               Method |      Mean |    Error |   StdDev |    Median |       Min |       Max | Gen 0 |     Gen 1 | Gen 2 | Allocated |
|--------------------- |----------:|---------:|---------:|----------:|----------:|----------:|-----------:|----------:|------:|----------:|
|  NormalizeWhitespace | 769.77 ms | 1.871 ms | 1.563 ms | 769.94 ms | 767.83 ms | 773.58 ms | 16000.0000 | 4000.0000 |     - |     94 MB |

Here is how I benchmarked (based on dotnet/performance):

```csharp
    [BenchmarkCategory("Roslyn")]
    public class RoslynApis
    {
        private SyntaxNode _root;

        [GlobalSetup]
        public void InitializeRoot()
        {
            // benchmark is run against a well-formatted file, not sure if it does matter.
            _root = SyntaxFactory.ParseCompilationUnit(File.ReadAllText(@"C:\Users\PC\Desktop\Roslyn\src\Compilers\CSharp\Portable\Generated\CSharpSyntaxGenerator\CSharpSyntaxGenerator.SourceGenerator\Syntax.xml.Internal.Generated.cs"));
        }

        [Benchmark]
        public SyntaxNode NormalizeWhitespace()
        {
            return _root.NormalizeWhitespace();
        }
    }
}
```

---

Another benchmark with a tree with no trivia (i.e, have the syntax normalizer do actual work and put trivia):

**Before:**

|              Method |     Mean |   Error |  StdDev |   Median |      Min |      Max |      Gen 0 |     Gen 1 | Gen 2 | Allocated |
|-------------------- |---------:|--------:|--------:|---------:|---------:|---------:|-----------:|----------:|------:|----------:|
| NormalizeWhitespace | 772.7 ms | 4.39 ms | 3.89 ms | 772.7 ms | 767.4 ms | 781.2 ms | 24000.0000 | 8000.0000 |     - |    140 MB |

**After:**

|              Method |     Mean |   Error |  StdDev |   Median |      Min |      Max |      Gen 0 |     Gen 1 | Gen 2 | Allocated |
|-------------------- |---------:|--------:|--------:|---------:|---------:|---------:|-----------:|----------:|------:|----------:|
| NormalizeWhitespace | 743.7 ms | 3.01 ms | 2.67 ms | 744.4 ms | 738.2 ms | 747.1 ms | 15000.0000 | 4000.0000 |     - |     89 MB |

Updated benchmark code:

```csharp
  public class TriviaRemover : CSharpSyntaxRewriter
    {
        public override SyntaxNode Visit(SyntaxNode node) => base.Visit(node)?.WithoutTrivia();

        public override SyntaxTrivia VisitTrivia(SyntaxTrivia trivia) => default;
    }

    [BenchmarkCategory("Roslyn")]
    public class RoslynApis
    {
        private SyntaxNode _root;

        [GlobalSetup]
        public void InitializeRoot()
        {
            _root = SyntaxFactory.ParseCompilationUnit(File.ReadAllText(@"C:\Users\PC\Desktop\Roslyn\src\Compilers\CSharp\Portable\Generated\CSharpSyntaxGenerator\CSharpSyntaxGenerator.SourceGenerator\Syntax.xml.Internal.Generated.cs"));
            _root = ((CSharpSyntaxNode)_root).Accept(new TriviaRemover());
        }

        [Benchmark]
        public SyntaxNode NormalizeWhitespace()
        {
            return _root.NormalizeWhitespace();
        }
    }
}
```

---

After commit 3 (introducing IsKind with multiple kinds) added an extra improvement:

|              Method |     Mean |   Error |  StdDev |   Median |      Min |      Max |      Gen 0 |     Gen 1 | Gen 2 | Allocated |
|-------------------- |---------:|--------:|--------:|---------:|---------:|---------:|-----------:|----------:|------:|----------:|
| NormalizeWhitespace | 723.8 ms | 4.16 ms | 3.48 ms | 723.4 ms | 717.5 ms | 729.8 ms | 15000.0000 | 4000.0000 |     - |     89 MB |

---