@cston 2021-07-04
What part of the change reduced allocations?