@Youssef1313 2021-07-04
> What part of the change reduced allocations?

I think none 😄 
Most likely this is due to benchmarking against 3.8.0 and my change, instead of main and change.