@Youssef1313 2021-07-04
I re-benchmarked `main` instead of 3.8.0, there doesn't seem to be any performance improvement. My expectation that this can improve performance originated from:

https://github.com/dotnet/roslyn-analyzers/blob/3e7a193ae1df08b77e368f02537a0d16c2648149/src/Microsoft.CodeAnalysis.Analyzers/Core/CodeAnalysisDiagnosticsResources.resx#L521, which indeed was convincing to me that `Kind()` does slightly more than `IsKind` (which just checks `RawKind`).

Also it's weird that I was seeing an improvement after commit 2 and commit 3 (i.e, this doesn't involve the mistake of benchmarking 3.8.0). But I can't now see the difference I was seeing previously.