@333fred 2021-07-09
> But I can't now see the difference I was seeing previously.

There was a perf-related PR after 3.8 from another contributor, it's possible that you were seeing the improvements from that PR.