@Youssef1313 2021-07-10
> > But I can't now see the difference I was seeing previously.
> 
> There was a perf-related PR after 3.8 from another contributor, it's possible that you were seeing the improvements from that PR.

Well, I'm going to close for now as this doesn't prove any perf gain (which I actually don't expect since IsKind should indeed be faster). However, it would be good to update the version used in dotnet/performance to the current latest version:

https://github.com/dotnet/performance/blob/main/src/benchmarks/real-world/Roslyn/CompilerBenchmarks.csproj#L18